requirements.txt
faster-whisper==1.0.3
numpy==1.24.3
moviepy==1.0.3
pyinstaller==6.3.0
PyAudioWPatch==0.2.12.8
scipy==1.11.4
